all: move model classes to database, rename model to api
