+++ .idea/gradle.xml
@@ -11,10 +11,10 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/api" />
             <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/buildSrc" />
             <option value="$PROJECT_DIR$/database" />
-            <option value="$PROJECT_DIR$/model" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
