+++ api/build.gradle.kts
@@ -9,8 +9,8 @@ dependencies {
   kapt(Dependencies.AndroidX.Hilt.daggerCompiler)
   kapt(Dependencies.ThirdParty.Moshi.codegen)
   api(Dependencies.ThirdParty.Retrofit.lib)
+  implementation(project(":database"))
   implementation(Dependencies.AndroidX.Hilt.dagger)
-  implementation(Dependencies.ThirdParty.Moshi.lib)
   implementation(Dependencies.ThirdParty.Retrofit.moshi)
   testImplementation(Dependencies.Kotlin.Coroutines.core)
   testImplementation(Dependencies.Testing.junit)
