+++ api/src/main/AndroidManifest.xml
