+++ api/src/main/java/dev/msfjarvis/lobsters/injection/BaseUrlQualifier.kt
