+++ database/src/main/java/dev/msfjarvis/lobsters/model/TagsAdapter.kt
@@ -1,0 +1,17 @@
+package dev.msfjarvis.lobsters.model
+
+import com.squareup.sqldelight.ColumnAdapter
+
+class TagsAdapter : ColumnAdapter<List<String>, String> {
+  override fun decode(databaseValue: String): List<String> {
+    return databaseValue.split(SEPARATOR)
+  }
+
+  override fun encode(value: List<String>): String {
+    return value.joinToString(SEPARATOR)
+  }
+
+  private companion object {
+    private const val SEPARATOR = ","
+  }
+}
